+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased](https://github.com/Instabug/Instabug-React-Native/compare/v15.01...dev)
+
+### Added
+
+- Add support for chaining errors . ([#1396](https://github.com/Instabug/Instabug-React-Native/pull/1396))
+
 ## [15.0.1](https://github.com/Instabug/Instabug-React-Native/compare/v14.3.0...v15.0.1)
 
 ### Added